+++ compiler/standard/src/main/java/gsonpath/Dependencies.kt
@@ -1,0 +1,8 @@
+package gsonpath
+
+import gsonpath.generator.adapter.AutoGsonAdapterGenerator
+import gsonpath.generator.factory.TypeAdapterFactoryGenerator
+
+data class Dependencies(
+        val autoGsonAdapterGenerator: AutoGsonAdapterGenerator,
+        val typeAdapterFactoryGenerator: TypeAdapterFactoryGenerator)