+++ compiler/standard/src/main/java/gsonpath/generator/TypeSpecExt.kt
@@ -2,25 +2,22 @@ package gsonpath.generator
 
 import com.squareup.javapoet.JavaFile
 import com.squareup.javapoet.TypeSpec
+import gsonpath.ProcessingException
 import gsonpath.util.FileWriter
-import gsonpath.util.Logger
 import java.io.IOException
 
 fun TypeSpec.Builder.writeFile(
         fileWriter: FileWriter,
-        logger: Logger,
         packageName: String,
-        fileBuiltFunc: (builder: JavaFile.Builder) -> Unit = {}): Boolean {
+        fileBuiltFunc: (builder: JavaFile.Builder) -> Unit = {}) {
 
-    return try {
+    try {
         JavaFile.builder(packageName, build()).apply {
             fileBuiltFunc(this)
             build().writeTo(fileWriter.filer)
         }
-        true
 
     } catch (e: IOException) {
-        logger.error("Error while writing javapoet file: " + e.message)
-        false
+        throw ProcessingException("Error while writing javapoet file: ${e.message}")
     }
 }