+++ compiler/testing/integration/src/test/resources/generator/extension/gson_sub_type/with_other_elements/TypesPojo_GsonTypeAdapter.java
@@ -108,7 +108,7 @@ public final class TypesList_GsonTypeAdapter extends TypeAdapter<TypesList> {
         String obj0 = value.other1;
         if (obj0 != null) {
             out.name("other1");
-            mGson.getAdapter(String.class).write(out, obj0);
+            writeWithGenericAdapter(mGson, obj0.getClass(), out, obj0);
         }
 
         Type obj1 = value.items;
@@ -121,7 +121,7 @@ public final class TypesList_GsonTypeAdapter extends TypeAdapter<TypesList> {
         String obj2 = value.other2;
         if (obj2 != null) {
             out.name("other2");
-            mGson.getAdapter(String.class).write(out, obj2);
+            writeWithGenericAdapter(mGson, obj2.getClass(), out, obj2);
         }
 
         // End