+++ sample/src/test/java/gsonpath/kotlin/SealedClassSampleTest.java
@@ -13,7 +13,7 @@ import java.io.InputStreamReader;
 public class SealedClassSampleTest {
     @Test
     public void testWithAllValues() {
-        SealedClassSample model = runTest("SealedClassSample.json");
+        SealedClassArray model = runTest(SealedClassArray.class, "SealedClassSample.json");
 
         Type item1 = model.getItems()[0];
         Assert.assertEquals("Type1 Example", item1.getName());
@@ -28,14 +28,42 @@ public class SealedClassSampleTest {
         Assert.assertEquals("123", ((Type.Type3) item3).getStringTest());
     }
 
-    private SealedClassSample runTest(String fileName) {
-        GsonBuilder builder = new GsonBuilder();
-        builder.registerTypeAdapterFactory(GsonPath.createTypeAdapterFactory(TestGsonTypeFactory.class));
+    @Test
+    public void testWrite() {
+        testWrite(new SealedClassPojo("name", new Type.Type1("Foobar", 1), "other"),
+                "{\"value1\":\"name\",\"item\":{\"common\":{\"name\":\"Foobar\"},\"specific\":{\"intTest\":1}},\"value2\":\"other\"}");
 
+        testWrite(new SealedClassPojo("name", new Type.Type2("Foobar", 1.5d), "other"),
+                "{\"value1\":\"name\",\"item\":{\"common\":{\"name\":\"Foobar\"},\"specific\":{\"doubleTest\":1.5}},\"value2\":\"other\"}");
+
+        testWrite(new SealedClassPojo("name", new Type.Type3("Foobar", "Sample"), "other"),
+                "{\"value1\":\"name\",\"item\":{\"common\":{\"name\":\"Foobar\"},\"specific\":{\"stringTest\":\"Sample\"}},\"value2\":\"other\"}");
+    }
+
+    private void testWrite(SealedClassPojo pojo, String json) {
+        Assert.assertEquals(json, createGson().toJson(pojo));
+    }
+
+    @Test
+    public void testSinglePojo() {
+        SealedClassSubTypePojo model = runTest(SealedClassSubTypePojo.class, "SealedClassSample.json");
+
+        Type item1 = model.getItem();
+        Assert.assertEquals("Type1 Example", item1.getName());
+        Assert.assertEquals(1, ((Type.Type1) item1).getIntTest());
+    }
+
+    private <T> T runTest(Class<T> clazz, String fileName) {
         ClassLoader classLoader = ClassLoader.getSystemClassLoader();
         InputStream resourceAsStream = classLoader.getResourceAsStream(fileName);
 
-        Gson gson = builder.create();
-        return gson.fromJson(new InputStreamReader(resourceAsStream), SealedClassSample.class);
+        return createGson().fromJson(new InputStreamReader(resourceAsStream), clazz);
+    }
+
+    private Gson createGson() {
+        GsonBuilder builder = new GsonBuilder();
+        builder.registerTypeAdapterFactory(GsonPath.createTypeAdapterFactory(TestGsonTypeFactory.class));
+
+        return builder.create();
     }
 }